Created StatsPage.tsx
