+++ client-app/src/getAzureFunctions.ts
@@ -48,8 +48,8 @@ const azureFunctionsProduction = {
     UserSignOut: "https://jebrafunctions.azurewebsites.net/api/UserSignOut?code=CbJQOhRTcgw7JoDMKH2y62ZZYe6AW3P18vkZX5iYACrhSwHYePBvlw=="
 };
 
-const getAzureFunctions = () => (process.env.NODE_ENV === "development") ? azureFunctionsDevelopment : azureFunctionsProduction;
+// const getAzureFunctions = () => (process.env.NODE_ENV === "development") ? azureFunctionsDevelopment : azureFunctionsProduction;
 // Comment the line above and uncomment the line below if you want to use prod endpoints in a dev environment (DO SO CAUTIOUSLY)
-//const getAzureFunctions = () => azureFunctionsProduction;
+const getAzureFunctions = () => azureFunctionsProduction;
 
 export default getAzureFunctions;
